Texture support with stb_image

+++ src/engine/engine.h
@@ -6,7 +6,7 @@
     x;             \
     _gl_call_impl(__FILE__, __LINE__)
 
-void _gl_call_impl(char *file, size_t line);
+void _gl_call_impl(char *file, int line);
 
 GLuint createAndCompileShader(const char *path, GLenum type);
